requirements.txt
confusable-homoglyphs==3.1.1
Django==1.11.13
django-registration==2.4.1
django-widget-tweaks==1.4.2
pytz==2018.4
